CHROMIUM: investigate reduction in hang detector timeout
For some undocumented (AFAIK) reason we set the hang detector timeout
to 120 seconds.  We'd like to keep it shorter because then fewer users
will long-press the power button, or reset the EC, before the timeout
expires.  This change produces a warning whenever a process is hung
for more than 60 seconds, and another one whenever a process that
was hung at 60 seconds becomes unhung.  By comparing the frequency
of these warnings we can tell if a shorter timeout is sufficient.

BUG=chromium:214618
TEST=none yet

Change-Id: I3ed1962401a4c43d62b08b8cf58369f17625a918
Signed-off-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/179333