CHROMIUM: i2c: samsung: resume the I2C controller earlier
When the wake-up is triggered by the PMIC RTC, the RTC driver is trying
to read the PMIC interrupt status over I2C and fails because the I2C
controller is not resumed yet.
Let's resume the I2C controller earlier in the _noirq phase
(as other hardwares are doing), so we can properly get the wake-up
condition.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chrome-os-partner:24427
TEST=On Spring, run power_Resume test and check that RTC alarm is still
working afterwards (simply run the test several times in a row)

Change-Id: I1cec05448ce6fb039acd90a4885b47c9ae3494a4
Reviewed-on: https://chromium-review.googlesource.com/179371
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>