ARM: dts: Cleanup peach pit dts files
We'd like to add more sharing between peach variants, so promote the
"peach-pit.dtsi" file to just a "peach.dtsi" file.  We'll then migrate
things back out to peach-pit as they're found not to be common.

While we're at it, also combine the pit-rev3 and pit-rev5 DTS files,
since the two boards have identical device tree.

BUG=chrome-os-partner:24492
TEST=Things still work fine on pit.

Change-Id: I09e7450cb87469e23121d9f78a6cb355e45047ed
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/179530
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>