CHROMIUM: ARM: init: Fix memory setup in the case of CONFIG_SPARSEMEM
The code in mem_init() was assuming that pages in a memory bank had
contiguous "struct page" entries.  That doesn't appear to always be
the case with CONFIG_SPARSEMEM.  Remove the slight optimization and
just calculate the location of the struct page each time.

BUG=chrome-os-partner:24537
TEST=Boot a kernel with an initramfs (which makes the kernel 6.7 MiB
big) and see that it no longer crashes.
TEST=Time boot performance and see no degredation.

Change-Id: I49142fc877b7e385c0a2a564d1f38f6bf268a125
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/180326
Reviewed-by: Olof Johansson <olofj@chromium.org>