mkbp: Fix problems with backslash
The mkbp code (post 3.4) has an issue where it can't deal with two
keys on the keyboard having the same keycode.  When this happens it
will get confused about whether the key is down or up and will cause
some screwy behavior.

We need to have two keys on the keyboard with the same keycode for UK
keyboards.  Specifically:
* On the US keyboard the backslash key (above enter) is r3 c11 and is
  supposed to be reported as BACKSLASH.
* On the UK keyboard the # key (left of enter) is r4 c10 and is
  supposed to be reported as BACKSLASH.
* On the UK keyboard the \ key (left of Z) is r2 c7 and is supposed to
  be reported as KEY_102ND.

The above may be crazy, but that's how it's supposed to work so we
need to handle two KEY_BACKSLASH keycodes.  Luckily this is easy to
handle.

BUG=chrome-os-partner:24379
TEST=Ran evtest and validated the above on US and UK keyboards.
TEST=Crazy double inputs of "\" and "|" when pressing shift and
backslash at the same time on US keyboard no longer happens.
TEST=Different repeat behavior when holding backslash and other keys
together on US keyboard no longer happens.

Change-Id: I175f335008fc2e6bb4118e9de013870e769b1df7
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/180494
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>