CHROMIUM: ARM: exynos: Add pmu dts nodes for Exynos 542x
This adds two pmu nodes to the exynos542x.dtsi because the two types
of cores unfortunately have different interrupt parents. This is just a
first step in getting performance counters to work properly on 5420
with IKS.  There still needs to be work done to support the mulitple
pmu nodes and then context switch the counters properly in the
switcher code.

BUG=chrome-os-partner:22742
TEST="for i in $(seq 0 3) ;
	  do echo 700000 > \
	  /sys/devices/system/cpu/cpu${i}/cpufreq/scaling_min_freq;
	  perf top works"

Change-Id: I720bf655a527e8e9e67608dd136c9cbd9aa1f6ef
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/180635
Reviewed-by: Olof Johansson <olofj@chromium.org>