exynos5420: devfreq: Disable 111 and 83 MHz levels for 1920x1080
Disables 83, 111MHz levels of devfreq for 1080p display
which comes with both of POP and SCP version of exynos5420.
This is required as the higher resolution demands higher
bus bandwidth.
This co-exists with the SCP code and applies only in case of 1080p or more.

BUG=chrome-os-partner:24554
TEST=Run Video, WebGL, V8 and see no stability issues.
Signed-off-by: Arjun.K.V <arjun.kv@samsung.com>

Change-Id: I9818bfc64b39a8fb249b272ea464adebd4167ff1
Reviewed-on: https://chromium-review.googlesource.com/180620
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Wonjoon Lee <woojoo.lee@samsung.com>
Tested-by: Wonjoon Lee <woojoo.lee@samsung.com>