drm/exynos: Set mixer powered flag to true earlier
This patch moves the "ctx->powered = true" assignment above interrupt enable
and window reset/resume in the mixer. Aside from being the Right
Thing To Do (so we correctly track the state of the hardware),
it is also necessary. If this is not set before we run enable_vblank,
we will defer the enable indefinitely by just setting int_en instead
of actually flipping the bit in MXR_INT_EN.

BUG=chromium:331863
TEST=Tested on daisy, external screen updates on replug/resume

Change-Id: I15069caaeddb043b4d18df9744824f6cdae952b9
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/181641
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>