drm: edid: enable probing and listing of non rb modes
The current solution checks for the existing RB mode,
if available in the edid block returns by adding it,
but does not populate the connector with the modes
of same resolution but which are non-rb modes.

As a result the probing and listing of non-rb modes can't
be made, in case the rb mode's pixel clock is not
supported but non-rb mode is supported.

This patch changes the drm_mode_std mode selection to
collect all the supported modes and not just one mode.

BUG=chrome-os-partner:24846
TEST=connect 1080p monitor to Daisy, 1680x1050 should be available

Change-Id: I0d36a374cea52ad7f118daed37b55f292f124fcf
Signed-off-by: Shirish S <s.shirish@samsung.com>
Signed-off-by: Dominik Behr <dbehr@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/181822
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>