CHROMIUM: usb: quirks: Make DELAY_INIT wait 100ms between GetConfiguration requests
The Logitech HD Pro Webcam C920 is a very quirky device: we could
reproduce all kinds of rare enumeration failures due to broken responses
to the second (full) GetConfigurationDescriptor request. The DELAY_INIT
quirk (which delays one second after the GetDeviceDescriptor request)
helps a bit but is not perfectly reliable.

This patch adds another 100ms delay time between the first and the
second GetConfigurationDescriptor request for DELAY_INIT devices. This
seems to reliably deal with the problems we encountered on the C920, and
shouldn't have a negative impact on other devices using this quirk.

BUG=chrome-os-partner:24893
TEST=Manual

We should eventually try to push this upstream, but I'd like to wait on
how reliably it really solves the issue and what Logitech turns up in
their root-cause investigation first.

Change-Id: I7cd77a66c76a20b0ea3fbfd8061c155788b9bd11
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/182702
Reviewed-by: Mohammed Habibulla <moch@google.com>
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>