CHROMIUM: ALSA: hda/ca0132 - setup/cleanup streams
When a HDMI stream is opened with the same stream tag
as a following opened stream to ca0132, audio will be
heard from two ports simultaneously.
Fix this issue by change to use snd_hda_codec_setup_stream
and snd_hda_codec_cleanup_stream instead, so that an
inactive stream can be marked as 'dirty' when found
with a conflict stream tag, and then get purified.

BUG=chromium:332201
TEST=On pixel, connect external monitor through HDMI:
1. play audio (to HDMI)
2. switch to internal speaker
3. play audio, and verify it comes only from intenral
speaker instead of both HDMI and speaker.

Change-Id: I5e9dc8c8ccdca47d47fc0aa9a02dec9dd7ab6d2f
Signed-off-by: Hsin-Yu Chao <hychao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/182994
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>