MALI: platform/chromebook: exynos5422: Add 543MHz support
Add support for 543MHz on Exynos5422. We were not using 533 MHz
on 5420 even though it was available because we were not hitting this
level for most benchmarks. The existent 5420 support can be re-used
without any changes for 5422 as the only difference between the two
is 533->543 but for now I have added a new table for 5422.

BUG=chrome-os-partner:25272
TEST=Build and boot on peach_pi. With a u-boot patch setting the
default rate to 420MHz do the follwing:

$ cd /sys
$ cat ./bus/platform/drivers/mali/11800000.mali/asv
asv group:4 mp6
543000000:975000
543000000:975000
480000000:925000
420000000:887500
350000000:862500
266000000:837500
177000000:775000
100000000:775000

$ cat ./bus/platform/drivers/mali/11800000.mali/clock
100000000

$ cat ./bus/platform/drivers/mali/11800000.mali/vol
Current operating voltage for mali t6xx = 775000

$ echo off > /sys/bus/platform/drivers/mali/11800000.mali/dvfs
G3D DVFS is disabled

$ cat /sys/bus/platform/drivers/mali/11800000.mali/clock
420000000

$ cat /sys/class/regulator/regulator.40/microvolts
887500

Change-Id: Ifc86d87a26b3d0597f3d493982eb43d0bf36f1a4
Signed-off-by: Abhilash Kesavan <a.kesavan@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/183415
Reviewed-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>