crypto:s5p-sss: Look for the next request in the queue
Currently, the driver enqueues a request only if the busy bit is
false. And every request initiates a dequeue. If 2 requests arrive
simultaneously, only one of them will be dequeued.

To avoid this senario, we will enqueue the next request irrespective
of the system condition (that is what queue is here for). Also
schedule at a tasklet immediatly after the current request is done.
The tasklet will dequeue the next request in the queue, giving
continuous loop. tasklet will exit if there are no requests in the
queue.

BUG=chrome-os-partner:23189, chrome-os-partner:25419
TEST=booted on peach_pit. login to see the /home/chronos/user/Downloads
being created. We should able to cp, rm, dd on to
/home/chronos/user/Downloads.

This patch set was reviewed at
http://lkml.org/lkml/2014/2/25/23

Change-Id: Iadc9fa80e65191cc77483c98051d6b90c4aa97f1
Signed-off-by: Naveen Krishna Ch <ch.naveen@samsung.com>
Acked-by: Herbert Xu <herbert@gondor.apana.org.au>
CC: David S. Miller <davem@davemloft.net>
CC: Herbert Xu <herbert@gondor.apana.org.au>
CC: Vladimir Zapolskiy <vzapolskiy@gmail.com>
TO: <linux-crypto@vger.kernel.org>
CC: <linux-samsung-soc@vger.kernel.org>
Reviewed-on: https://chromium-review.googlesource.com/184279
Reviewed-by: naveen krishna <naveen@chromium.org>
Tested-by: naveen krishna <naveen@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: Sonny Rao <sonnyrao@chromium.org>