mfd: tps65090: Stop caching registers
Nearly all of the registers in tps65090 combine control bits and
status bits.  Turn off caching of registers so that we can read status
bits reliably.

NOTE: the IRQnMASK and CG_CTRLn registers are the exception and could
be cached.  If we find that we spend a lot of time reading those we
can turn on cache for just those registers.

BUG=chromium:338657
TEST=Future changes that check tps65090 status now work.

Change-Id: I1850d7bb209a73c2cbe216289c0fe61be2252350
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/184606
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>