UPSTREAM: perf/x86/uncore: fix initialization of cpumask
On certain processors, the uncore PMU boxes may only be
msr-bsed or PCI-based. But in both cases, the cpumask,
suggesting on which CPUs to monitor to get full coverage
of the particular PMU, must be created.

However with the current code base, the cpumask was only
created on processor which had at least one MSR-based
uncore PMU. This patch removes that restriction and
ensures the cpumask is created even when there is no
msr-based PMU. For instance, on SNB client where only
a PCI-based memory controller PMU is supported.

Cc: mingo@elte.hu
Cc: acme@redhat.com
Cc: ak@linux.intel.com
Cc: zheng.z.yan@intel.com
Cc: peterz@infradead.org
BUG=chromium:474724
TEST=build/boot on Link

Change-Id: Iee3fa49e1d63da92404f09989a9a0da0abd895ff
Signed-off-by: Stephane Eranian <eranian@google.com>
Signed-off-by: Peter Zijlstra <peterz@infradead.org>
Link: http://lkml.kernel.org/r/1392132015-14521-2-git-send-email-eranian@google.com
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
(cherry picked from commit 411cf180fa00521f9bfb1d022e3ebf059a2d299f)
Reviewed-on: https://chromium-review.googlesource.com/264481
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>