UPSTREAM: perf/x86/uncore: make hrtimer timeout configurable per box
This patch makes the hrtimer timeout configurable per PMU
box. Not all counters have necessarily the same width and
rate, thus the default timeout of 60s may need to be adjusted.

This patch adds box->hrtimer_duration. It is set to default
when the box is allocated. It can be overriden when the box
is initialized.

Cc: mingo@elte.hu
Cc: acme@redhat.com
Cc: ak@linux.intel.com
Cc: zheng.z.yan@intel.com
Cc: peterz@infradead.org
BUG=chromium:474724
TEST=build/boot on Link

Change-Id: Iabed20bd9c0f7523503390285ac6a7385af8c85c
Signed-off-by: Stephane Eranian <eranian@google.com>
Signed-off-by: Peter Zijlstra <peterz@infradead.org>
Link: http://lkml.kernel.org/r/1392132015-14521-5-git-send-email-eranian@google.com
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
(cherry picked from commit 79859cce5a551c3bcd8d36a6287f9d83568ccedb)
Reviewed-on: https://chromium-review.googlesource.com/264483
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>