CHROMIUM: mfd: cros_ec: don't throttle in ec_dev_lightbar_init()
During init, we ask the EC for the lightbar version as a way to detect its
presence. That query shouldn't be throttled. Only the userspace interface
needs to be throttled to prevent DoS.

BUG=chromium:239205
BRANCH=none
TEST=autotest coming soon: crbug.com/343270

I haven't proven that this causes a problem, but it conceivably could. Might
as well avoid it.

Change-Id: I606dc59c0095c0122a3b7706eae495dc1de04021
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/186363
Tested-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>