drm/exynos: restore core HDMI settings
In DVI mode the video preamble and Guard band should
be disabled whereas it should be applied in HDMI mode,
the re-applying of preamble and guard band was missing,
which resulted in display failures when switched to HDMI
mode from DVI mode.
This patch ensures the setting is applied in HDMI mode.

BUG=chrome-os-partner:25762
TEST=connect 1080p monitor to Pit via DVI connector,
    unplugged and connected to HDMI TV via HDMI cable,
    and verifed ui on HDMI TV and vice versa.

Change-Id: I280858e99142fd293e3cbec30ad01e251d80db5c
Signed-off-by: Shirish S <s.shirish@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/186338
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Sean Paul <seanpaul@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>