drm/anx7808: Introduce the big lock
This patch adds a new big lock which replaces the previous mutexes. This
big lock is meant to keep modeset/detect/get_modes/cable_det/intp all
separate and mutually exclusive. The hope is that this will eliminate
the edge cases where pathological plugging/unplugging can get the device
in a bad state.

BUG=chromium:341790
TEST=On spring, rapidly plug/unplug dongle, screen should come up

Change-Id: I6573389c065c1a7bb789df212af04266df10c24f
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/186882