ARM: dts: Assign FET5 as the vcc for the light sensor
We'd like FET5 to get powered off while the device is suspended.
Assign it as the vcc for the light sensor so it will power on at
bootup and power off during suspend.

NOTE: Technically FET5 actually powers on _both_ the camera and the
light sensor, so we really ought to assign this as vcc to both
components.  Unfortunately there's no clean place to list the camera's
vcc since it's a USB device.  However, having the camera lose power at
the exact same time as the light sensor isn't terrible, so we'll just
do that.

BUG=chrome-os-partner:26000
TEST=On pit EC console type "pmu" while suspended and see 0x13=0x02
TEST=No problems / warnings booting snow/pit.
TEST=No problems / warnings suspend_stress_test on snow/pit.
TEST=Light sensor works on snow/pit before/after s2r.
TEST=Camera works on snow/pit before/after s2r.

Change-Id: I5fb26191a14655d8397de156db5a20b20cf6f1c9
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/187114
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>