CHROMIUM: config: enable Quickstep driver as module by default
The Quickstep driver was added to kernel-next in a previous patch, but
by default is not compiled into the kernel.  Since we'd like our test
devices to have it if a Quickstep device is plugged into them, but most
of the time it won't be getting used, this CL updates the hid Kconfig to
mark the Quickstep driver to be included as a module.

BUG=chromium:328451
TEST=compiled the kernel and the driver was there and functional

Change-Id: Ia346a2a16a41b81136a697614e15e32fc218880b
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/187270
Reviewed-by: Olof Johansson <olofj@chromium.org>