BACKPORT: mwifiex: Track BA sequence number reset
Some stations reset the sequence number for traffic-ids (TIDs)
as they initiate a block-ACK session.  In order to detect such
behavior, mwifiex must note the starting sequence number given
during the ADDBA request.  If the first received sequence number
after the ADDBA falls outside the receive window for this TID but
after the the ADDBA starting sequence number, we can assume that
this AP has reset its sequence number during the ADDBA.  In this
case we must adjust the input window backward to incorporate this
received sequence number instead of ignoring it.  Otherwise, we
could fail to successfully retrieve an arbitrarily large number
of downstream frames at the beginning of the block-ACK session.

Signed-off-by: Paul Stewart <pstew@chromium.org>

Originally-Reviewed-on: https://chromium-review.googlesource.com/187783
Backported to both 3.8 native and 3.8-with-3.4-wlan.

BUG=chrome-os-partner:25630,chromium:297607
TEST=Connect to 802.11n WPA/CCMP AP which resets its TIDs and
ensure initial DHCP response is delivered to userspace.

Change-Id: Ib447779d635327ea19e8b7ca0361e1a8e5e62b00
Reviewed-on: https://chromium-review.googlesource.com/187937
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>