arm: exynos: pm_domains: reconfigure dout clock during pd power on
The CLKDIV2_BLK for various ips is a mirror register of the actual register
inside the ip's domain, which gets reset when the power domain is turned off.
To set this internal register, the divider register needs to be reconfigured
(i.e. first reset and then set again). The dout clocks of CLKDIV2 block are
passed via the device node.

During each power on, if this dclk is present, the clock rate of this
clock is first set to its parent's clock rate (which resets the divider)
and then set back to the original dclk rate. This is done when the parent
clock is still the oscillator clock so we don't have to worry about over
clocking.

BUG=chrome-os-partner:26088
TEST=Play 1080p video on pi and pit for mfc.
Run veau test app for gsc on pit and pi.

Change-Id: I01ea4b5681c05a791edec75d8edffbde0a9e1a19
Signed-off-by: Prathyush K <prathyush.k@samsung.com>
Signed-off-by: Arun Mankuzhi <arun.m@samsung.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/187589
Tested-by: Arun Mankuzhi <arun@chromium.org>