CHROMIUM: pstore: Move to larger record size (again)
Accidentally specified a smaller record size than before, bring it back
to the same size as we had when we used the config file.

This was already done in CL 170128, but the switch over to hardcoded values
was based on earlier code so the change was lost in the move over.

BUG=chromium:295127
TEST=none

Change-Id: Ie3c73d90a4f7a7588f3b9377d16ea73e4170ff37
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/188076
Reviewed-by: Sameer Nanda <snanda@chromium.org>