mmc: dw_mmc: Make sure we don't get stuck when we get an error
If we happened to get a data error at just the wrong time the dw_mmc
driver could get into a state where it would never complete its
request.  That would leave the caller just hanging there.

We fix this two ways and both of the two fixes on their own appear to
fix the problems we've seen:

1. Fix a race in the tasklet where the interrupt setting the data
   error happens _just after_ we check for it, then we get a
   EVENT_XFER_COMPLETE.  We fix this by repeating a bit of code.
2. Fix it so that if we detect that we've got an error in the "data
   busy" state and we're not going to do anything else we end the
   request and unblock anyone waiting.

BUG=chrome-os-partner:26256
BUG=chrome-os-partner:26473
TEST=suspend_stress_test works reliably with hs200 now
TEST=See I27c95590907e5db8e6402719e73ff57ee0cc3fd5

Change-Id: Ie8fea8281ebbcdbab4a95181817a6961a5cac976
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/189141
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>