drm/anx7808: Implement mode_valid for anx7808 bridge.
This patch implements mode_valid for the anx7808 bridge by passing
the request upstream to the encoder that the bridge was initialized
with.  This allows modes that need to be rejected during the
drm_helper_probe_single_connector_modes preparation to be (that
normally would be rejected by the connector).

BUG=chromium:347700
TEST=Tested on spring with HDMI

Change-Id: I4c74b5162484b123744871f1f9f21fb324bc6e09
Signed-off-by: David Riley <davidriley@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/188249
Reviewed-by: Sean Paul <seanpaul@chromium.org>