arm: asv: exynos542x: Do not modify MIF and MIF_SRAM voltages
This patch retains MIF and MIF_SRAM voltages from u-boot. This is
based on the following observations from the hardware team:

The default power-on MIF voltage is 1.0V, and DRAM init is performed
at this voltage level, 1.0V.
That means that all the trainings like gate training and read DQ
training are done at 1.0V to find the optimal timing. However, in the
kernel, VDD_MIF is changed to each silicon's own ASV voltage, and it
shakes the optimal timing because the voltage change affects DLL lock
value and IO timing subsequently.

Therefore, timing can be worse in the kernel after setting to ASV
voltage level.
	-> To resolve it, there're two options.
	1) Perform the training at each ASV voltage.
	2) Set the operating voltage to the same level as the training voltage.

We are implementing 2) considering the change would otherwise have to be
done in SPL which in turn would require something like this:
https://chromium-review.googlesource.com/#/c/175267/

BUG=chrome-os-partner:26193
TEST=Stress test 10 boards overnight with memtester and observe no
failures.

Change-Id: I3d123525812adc6d2175297eeb90f6699bdb72b9
Signed-off-by: Abhilash Kesavan <a.kesavan@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/188470
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>