devfreq: exynos5422: Enable all levels, but keep voltage at highest level
This patch keeps INT locked to highest voltage and not frequency.
This exercises the whole of devfreq levels from 83-500MHz.
Also tunes the table so as to keep the "core clocks" running
at a reasonably high level so that we dont need locking
for HDMI/MMC/Video playback.

BUG=chrome-os-partner:26245
TEST=Boot on Pi. Run V8, WebGL, Video and run power_LoadTest.

Change-Id: I80170cc9939552d4015ef3ec50946a65232b8758
Signed-off-by: Arjun.K.V <arjun.kv@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/188523
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: naveen krishna <naveen@chromium.org>
Tested-by: naveen krishna <naveen@chromium.org>