CHROMIUM: usb: ehci-s5p: Conditionally reset HSIC phy for different boards
The HSIC phy is being reset on-the-fly unconditionally for all the
boards using ehci-s5p controller. This workaround however, should be
board dependent, so that only the boards which require, enable it.
So adding a board specific DT property for the same.
Moreover we have seen the HSIC phy reset mechanism behavior different
on different boards.
On snow things are pretty much fine, but with peach-pit and peach-pi,
a reset on HSIC phy makes the port to misbehave and thereby stopping
the 3503 hub to get detected.
Also removing the config USB_EHCI_S5P_HSIC_RESET, since we won't
require it if we have a board specific DT property.

BUG=chrome-os-partner:25212, chrome-os-partner:18932, chrome-os-partner:21952
TEST=Test on snow, peach-pit, and peach-pi; and other exynos5 based
	systems, and confirm that the HSIC hub (0424:3503) gets detected
	on them. See below result of lsusb on snow.
	localhost ~ # lsusb
	Bus 003 Device 002: ID 0424:3503 Standard Microsystems Corp.
	Bus 001 Device 001: ID 1d6b:0002 Linux Foundation 2.0 root hub
	Bus 002 Device 001: ID 1d6b:0003 Linux Foundation 3.0 root hub
	Bus 003 Device 001: ID 1d6b:0002 Linux Foundation 2.0 root hub
	Bus 004 Device 001: ID 1d6b:0001 Linux Foundation 1.1 root hub
	Bus 003 Device 003: ID 2232:1037

Change-Id: I3b2e5e3644e02461a01e78ad2f1a8d50eafc766c
Signed-off-by: Vivek Gautam <gautam.vivek@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/188999
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>