UPSTREAM: r8152: support IPv6
Support hw IPv6 checksum for TCP and UDP packets.

Note that the hw has the limitation of the range of the transport
offset. Besides, the TCP Pseudo Header of the IPv6 TSO of the hw
bases on the Microsoft document which excludes the packet length.

Signed-off-by: Hayes Wang <hayeswang@realtek.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 6128d1bb30748d0ff56a63898d14f312126e404c)

BUG=chrome-os-partner:24570
TEST=manual testing with RealTek USB rtl8153

Change-Id: I40e0061152d252a05adf8610d03e7b4868e05c0d
Signed-off-by: Inki Yoo <inki.yoo@samsung.com>
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/192281