ARM: Don't ever downscale loops_per_jiffy in SMP systems
Downscaling loops_per_jiffy on SMP ARM systems really doesn't work.
You could really only do this if:

* Each CPU is has independent frequency changes (changing one CPU
  doesn't affect another).
* We change the generic ARM udelay() code to actually look at percpu
  loops_per_jiffy.

I don't know of any ARM CPUs that are totally independent that don't
just use a timer-based delay anyway.  For those that don't have a
timer-based delay, we should be conservative and overestimate
loops_per_jiffy.

Note that on snow you sometimes see (in the extreme case when we're
all the way downclocked) a udelay(100) become a udelay(1000) now.
That's no worse than we were in kernel-3.4.  On pit/pi we don't
downscale our A7s as much so we see worse case more like 100=>300.

BUG=chrome-os-partner:26000
TEST=No more SPI transfer timeouts with the EC (with next CL, too).
TEST=See I656b3ec5810920c57e17fe32930d35926e3cc314 (on snow, pit, pi)

Change-Id: I33fbfc59a93deaeec096408b1ef46295de56f495
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/189885
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>