CHROMIUM: Bluetooth: Add intel_btusb module for Intel Bluetooth device
Intel Bluetooth device (7260) requires device firmware patching before
the host send HCI initialization commands. This setup routine was added
in 3.10 kernel with a framework support from the Bluetooth subsystem
which is also introduced in 3.10 kernel.

Because the backporting this setup routine from 3.10 to 3.8 requires
also backporting Bluetooth framework change, which is significant in
the Bluetooth subsystem, it is not recommened by the Google maintainer.

Thus, intel_btusb module is forked from btusb module to have its own
device firmware setup routine without changing Bluetooth susbsystem.

Even if it is for Intel Bluetooth device, any other routines such as
submit USB urb, interrupt handling, and ISOC setup are same as btusb
module.

This is a custom patch for ChromeOS

BUG=chrome-os-partner:21770
TEST=tested with HID mouse and A2DP headset

Change-Id: Ic62d6794fb39b23db4ce31a303c04d9b5285ee8b
Reviewed-on: https://chromium-review.googlesource.com/190552
Reviewed-by: Scott James Remnant <keybuk@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Queue: Tedd An <tedd.an@intel.com>
Tested-by: Tedd An <tedd.an@intel.com>